Added a skeleton project proposal.
This code was originally intended for the 2013 robosub competition,
but it was never considered for inclusion. I redesigned the project
from the bottom up to modularize the project. I created files intended
for the 2013 tasks, but we won't know what the 2014 tasks are for
several months anyway.

This code does not stub out any communication protocols or
microcontroller code.