Moved settings and changing module names.
This moves the information that was in settings.json to
robosub_settings.py. The reason for this change is that we can now
comment inside of the settings file. I picked the name
"robosub_settings.py" because it is more meaningful when it's imported
from various modules.

I only changed names of files that are references from
robosub_settings.py. Other modules can be modified more easily locally.

I fixed the spelling s/stabalization/stabilization

I also removed the movement/physical folder. The functionality that was
supposed to live there is now in the "microcontroller" folder.

The jumpstart program "Dr. Frankenstein's Machine" will now launch all
modules, with the exception of the keyboard input.