Renamed microcontroller interface files, started adding debugging interface files.
Also, renamed microcontroller_interface_LEDupdate to LED_microcontroller_interface, because that is what it actually is.

I renamed the files so that they will be easier to tab-autocomplete when calling them. The debugging files are not in a working state, they have just been started.

deleted logfile (not supposed to be in git)